We have static files served on S3!
Incredibly frustrating and spent maybe 12 hours over a few days trying to figure out what was wrong. It wasn't my code. It was some obscure thing on AWS S3 that needed to make all of the files public. Now, before launch I will need to find a way to make them a bit more secure.

The settings in this comit work perfectly, though. Going to try adding the compress back in.